Modeling: Fix logits processors (probs, biasing, lua)
